Refactor: LeetCode component and API hooks
