+++ popup/src/lib/leetpush.api.ts
@@ -4,14 +4,15 @@ import {
   UserStreakI,
 } from "@/types/leetpush.interface.ts";
 
+const BASE_URL =
+  "https://leet-push-api-git-master-husamahmuds-projects.vercel.app/api/v2";
+
 /**
  * fetchDailyProblem - Fetch the daily problem from the LeetPush API
  * @returns The daily problem
  **/
 export const fetchDailyProblem = async (): Promise<DailyProblemI> => {
-  const response = await fetch(
-    `https://leet-push-api-git-master-husamahmuds-projects.vercel.app/api/v2/daily`,
-  );
+  const response = await fetch(`${BASE_URL}/daily`);
   if (!response.ok) throw new Error("Failed to fetch the daily problem");
 
   const data = await response.json();
@@ -24,9 +25,7 @@ export const fetchDailyProblem = async (): Promise<DailyProblemI> => {
  * @returns The user stats
  **/
 export const fetchUserStats = async (username: string): Promise<UserStatsI> => {
-  const response = await fetch(
-    `https://leet-push-api-git-master-husamahmuds-projects.vercel.app/api/v2/${username}`,
-  );
+  const response = await fetch(`${BASE_URL}/${username}`);
   if (response.status === 404) throw new Error("User not found");
   else if (!response.ok) throw new Error("Failed to fetch user stats");
 
@@ -41,9 +40,7 @@ export const fetchUserStats = async (username: string): Promise<UserStatsI> => {
 export const fetchUserStreak = async (
   username: string,
 ): Promise<UserStreakI> => {
-  const response = await fetch(
-    `https://leet-push-api-git-master-husamahmuds-projects.vercel.app/api/v2/userProfileCalendar/${username}`,
-  );
+  const response = await fetch(`${BASE_URL}/userProfileCalendar/${username}`);
   if (!response.ok) throw new Error("Failed to fetch user streak");
 
   return await response.json();
